requirements.txt
requests
pinecone-client
firecrawl
python-dotenv
llama-index
llama-index-vector-stores-pinecone
llama-index-embeddings-openai
tiktoken

nltk
scipy

langchain
langchain_community
langchain_openai
# llama_index.llms.langchain

#backend
fastapi
uvicorn

#frontend
streamlit
streamlit_chat